// src/components/buttons/buttons.interface.tsx
export type ButtonColorType =
  | "inherit"
  | "primary"
  | "secondary"
  | "success"
  | "error"
  | "warning"
  | "info";

export type ButtonVariantType = "text" | "outlined" | "contained";

export type ButtonSizeType = "small" | "medium" | "large";

export type ButtonTypeType = "submit" | "button" | "reset";
// ...
export interface CommonButtonPropsI {
  children: ReactNode;
  onClick?: (...args: any) => void;
  variant?: ButtonVariantType;
  color?: ButtonColorType;
  disabled?: boolean;
  className?: string;
  fullWidth?: boolean;
  customStyles?: any;
  type?: ButtonTypeType;
  primary?: boolean;
  startIcon?: ReactNode | undefined;
  endIcon?: ReactNode | undefined;
}